pyproject.toml
[project]
name = "minesweepersolver"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mss>=10.1.0",
    "numpy>=2.4.4",
    "opencv-python>=4.13.0.92",
    "pillow>=12.2.0",
    "pyautogui>=0.9.54",
]
